[ConstraintSystem] NFC: Move `diagnoseAmbiguity` methods from `CSDiag` to ConstraintSystem`
